split off TestMultiIndexSlicers class
